Levels: Update C047L001.DAT and add C047L003.DAT ReLevelNo for Deep Mind
Restores a dummied out cut scene that occurs at the start of the level showing some Elite Zealots dropping off the kidnapped professor Drennan at their base. A new version of the level C047L003.DAT is now used as a ReLevelNo to skip the cutscene. This also has the start camera angle amended to what is the final angle after the cut scene, as this better fits the actual gameplay.